MainPage MVVM
